requirements.txt
streamlit>=1.28.0
langchain>=0.2.0
langchain-community>=0.0.10
boto3>=1.34.0
chromadb>=0.4.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
